+++ Pulsus/Gameplay/Event/MeasureMarkerEvent.cs
@@ -5,7 +5,7 @@ namespace Pulsus.Gameplay
 	[DebuggerDisplay("{ToString()}")]
 	public class MeasureMarkerEvent : Event
 	{
-		public MeasureMarkerEvent(int pulse)
+		public MeasureMarkerEvent(long pulse)
 			: base(pulse)
 		{
